python-etl/requirements.txt
apache-airflow-providers-amazon==9.2.0
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
wheel==0.45.1
fastapi
uvicorn
python-jose[cryptography]
requests

